requirements.txt
asgiref==3.8.1; python_version >= '3.8'
django==3.1.14; python_version >= '3.6'
djangorestframework==3.11.2; python_version >= '3.5'
psycopg2-binary==2.9.9; python_version >= '3.7'
pytz==2024.1
setuptools==69.2.0; python_version >= '3.8'
sqlparse==0.4.4; python_version >= '3.5'
prometheus_client==0.20.0; python_version >= '3.7'
django-prometheus==2.3.1